@@ -1,7 +1,5 @@
 """cogent3 apps for piqtree2."""
 
-from typing import Optional, Union
-
 import cogent3
 import cogent3.app.typing as c3_types
 from cogent3.app.composable import define_app
@@ -14,21 +12,21 @@ from piqtree2.model import Model
 @define_app
 @extend_docstring_from(build_tree)
 def piqtree_phylo(
-    aln: Union[cogent3.Alignment, cogent3.ArrayAlignment],
+    aln: cogent3.Alignment | cogent3.ArrayAlignment,
     model: Model,
-    rand_seed: Optional[int] = None,
-) -> Union[cogent3.PhyloNode, cogent3.app.typing.SerialisableType]:
+    rand_seed: int | None = None,
+) -> cogent3.PhyloNode | cogent3.app.typing.SerialisableType:
     return build_tree(aln, model, rand_seed)
 
 
 @define_app
 @extend_docstring_from(fit_tree)
 def piqtree_fit(
-    aln: Union[cogent3.Alignment, cogent3.ArrayAlignment],
+    aln: cogent3.Alignment | cogent3.ArrayAlignment,
     tree: cogent3.PhyloNode,
     model: Model,
-    rand_seed: Optional[int] = None,
-) -> Union[cogent3.PhyloNode, cogent3.app.typing.SerialisableType]:
+    rand_seed: int | None = None,
+) -> cogent3.PhyloNode | cogent3.app.typing.SerialisableType:
     return fit_tree(aln, tree, model, rand_seed)
 
 
@@ -38,7 +36,7 @@ def piqtree_random_trees(
     num_taxa: int,
     tree_mode: TreeGenMode,
     num_trees: int,
-    rand_seed: Optional[int] = None,
+    rand_seed: int | None = None,
 ) -> tuple[cogent3.PhyloNode]:
     return random_trees(num_taxa, tree_mode, num_trees, rand_seed)
 
