@@ -1,7 +1,7 @@
 # testing the display of functions
-from typing import Optional
 
 import pytest
+
 from piqtree2 import available_freq_type, available_models, available_rate_type
 from piqtree2.model import AaModel, DnaModel, FreqType, SubstitutionModel
 from piqtree2.model._rate_type import ALL_BASE_RATE_TYPES
@@ -12,8 +12,8 @@ from piqtree2.model._rate_type import ALL_BASE_RATE_TYPES
     [(None, None), (DnaModel, "dna"), (AaModel, "protein")],
 )
 def test_num_available_models(
-    model_class: Optional[SubstitutionModel],
-    model_type: Optional[str],
+    model_class: SubstitutionModel | None,
+    model_type: str | None,
 ) -> None:
     table = available_models(model_type)
     total_models = (
@@ -28,8 +28,8 @@ def test_num_available_models(
     [(None, None), ("dna", "nucleotide"), ("protein", "protein")],
 )
 def test_available_models_types(
-    model_fetch: Optional[str],
-    model_type: Optional[str],
+    model_fetch: str | None,
+    model_type: str | None,
 ) -> None:
     table = available_models(model_fetch)
 
